+++ rsvim_core/src/ui/canvas/internal.rs
@@ -1,3 +1,6 @@
 //! Internal implementations for `Frame`.
 
 pub mod iframe;
+
+#[cfg(test)]
+mod iframe_tests;
